refactor(dapp): reduce duplication and normalize responses
